move otel extensions to separate file
