+++ src/Dafda/Diagnostics/OpenTelemetryMessagingAttributes.cs
@@ -1,9 +1,13 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Reflection;
+using Confluent.Kafka;
 using Dafda.Consuming;
+using Dafda.Producing;
+using Dafda.Serializing;
 using OpenTelemetry;
 using OpenTelemetry.Context.Propagation;
+using Metadata = Dafda.Consuming.Metadata;
 
 namespace Dafda.Diagnostics;
 
@@ -39,11 +43,6 @@ internal static class OpenTelemetryMessagingAttributes
     public const string OPERATION = "messaging.operation";
 
     /// <summary>
-    /// The conversation ID identifying the conversation to which the message belongs, represented as a string. Sometimes called “Correlation ID”.
-    /// </summary>
-    public const string CONVERSATION_ID = "messaging.message.conversation_id";
-
-    /// <summary>
     /// A unique identifier for the client that consumes or produces a message.
     /// </summary>
     public const string CLIENT_ID = "messaging.client_id";
@@ -62,58 +61,6 @@ internal static class OpenTelemetryMessagingAttributes
     /// Name of the Kafka Consumer Group that is handling the message. Only applies to consumers, not producers.
     /// </summary>
     public const string KAFKA_CONSUMER_GROUP = "messaging.kafka.consumer.group";
-}
-
-internal static class OpenTelemetryActivityExtensions
-{
-    private static class Messaging
-    {
-        public const string System = "kafka";
-        public const string DestinationKind = "topic";
-    }
-
-    private static class Operation
-    {
-        public const string Receive = "receive";
-        public const string Publish = "publish";
-    }
-
-    public static Activity AddDefaultOpenTelemetryTags(this Activity activity,
-        string topicName,
-        string messageId,
-        string conversationId,
-        string clientId,
-        string partitionKey,
-        int partition)
-    {
-        // messaging tags
-        activity.SetTag(OpenTelemetryMessagingAttributes.SYSTEM, Messaging.System);
-        activity.SetTag(OpenTelemetryMessagingAttributes.DESTINATION_KIND, Messaging.DestinationKind);
-
-        activity.SetTag(OpenTelemetryMessagingAttributes.DESTINATION, topicName);
-        activity.SetTag(OpenTelemetryMessagingAttributes.MESSAGE_ID, messageId);
-        activity.SetTag(OpenTelemetryMessagingAttributes.CONVERSATION_ID, conversationId);
-        activity.SetTag(OpenTelemetryMessagingAttributes.CLIENT_ID, clientId);
-
-        // kafka specific tags
-        activity.SetTag(OpenTelemetryMessagingAttributes.KAFKA_MESSAGE_KEY, partitionKey);
-        activity.SetTag(OpenTelemetryMessagingAttributes.KAFKA_PARTITION, partition);
-
-        return activity;
-    }
-
-    public static Activity AddConsumerOpenTelemetryTags(this Activity activity, string groupId)
-    {
-        activity.SetTag(OpenTelemetryMessagingAttributes.KAFKA_CONSUMER_GROUP, groupId);
-        activity.SetTag(OpenTelemetryMessagingAttributes.OPERATION, Operation.Receive);
-        return activity;
-    }
-
-    public static Activity AddProducerOpenTelemetryTags(this Activity activity)
-    {
-        activity.SetTag(OpenTelemetryMessagingAttributes.OPERATION, Operation.Publish);
-        return activity;
-    }
 }
 
 internal static class ConsumerActivitySource
@@ -138,7 +85,6 @@ internal static class ConsumerActivitySource
             .AddDefaultOpenTelemetryTags(
                 topicName: @event.Topic,
                 messageId: @event.Message.Metadata.MessageId,
-                conversationId: @event.Message.Metadata.CorrelationId,
                 clientId: @event.ClientId,
                 partitionKey: @event.PartitionKey,
                 partition: @event.Partition)
@@ -149,5 +95,37 @@ internal static class ConsumerActivitySource
     private static IEnumerable<string> ExtractFromMetadata(Metadata metadata, string key)
     {
         yield return metadata[key];
+    }
+}
+
+internal static class ProducerActivitySource
+{
+    private const string ActivityNameSuffix = "send";
+    private static readonly AssemblyName AssemblyName = typeof(KafkaProducer).Assembly.GetName();
+    private static readonly ActivitySource ActivitySource = new(AssemblyName.Name, AssemblyName.Version.ToString());
+    public static TextMapPropagator Propagator { get; set; } = Propagators.DefaultTextMapPropagator;
+
+    public static Activity StartActivity(PayloadDescriptor payloadDescriptor)
+    {
+        // Extract the current activity context
+        var contextToInject = Activity.Current?.Context
+                              ?? default;
+
+        // Inject the current context into the message headers
+        Propagator.Inject(new PropagationContext(contextToInject, Baggage.Current), payloadDescriptor, InjectTraceContext);
+
+        // Start the activity
+        return ActivitySource.StartActivity($"{payloadDescriptor.TopicName} {ActivityNameSuffix}", ActivityKind.Producer)
+            .AddDefaultOpenTelemetryTags(
+                topicName: payloadDescriptor.TopicName,
+                messageId: payloadDescriptor.MessageId,
+                clientId: payloadDescriptor.ClientId,
+                partitionKey: payloadDescriptor.PartitionKey)
+            .AddProducerOpenTelemetryTags();
+    }
+
+    private static void InjectTraceContext(PayloadDescriptor descriptor, string key, string value)
+    {
+        descriptor.AddHeader(key, value);
     }
 }
